feat: add bandwith support for knn density estimation.
