+++ src/density/knn.rs
@@ -1,30 +1,10 @@
 use core::f64::consts::PI;
 
-use kdtree::{distance::squared_euclidean, ErrorKind, KdTree};
-use thiserror::Error;
+use kdtree::{distance::squared_euclidean, KdTree};
 
-use crate::function::gamma::gamma;
+use crate::{density::DensityError, function::gamma::gamma};
 
 use super::Container;
-
-#[derive(Error, Debug)]
-pub enum DensityError {
-    /// Error when the k-d tree cannot be built or queried.
-    #[error(transparent)]
-    KdTree(#[from] ErrorKind),
-    EmptySample,
-    EmptyNeighborhood,
-}
-
-impl core::fmt::Display for DensityError {
-    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-        match self {
-            DensityError::KdTree(err) => write!(f, "K-d tree error: {}", err),
-            DensityError::EmptySample => write!(f, "No samples provided"),
-            DensityError::EmptyNeighborhood => write!(f, "No neighbors found"),
-        }
-    }
-}
 
 fn orava_optimal_k(n_samples: f64) -> f64 {
     // Adapted from K-nearest neighbour kernel density estimation, the choice of optimal k; Jan Orava 2012
@@ -37,7 +17,7 @@ fn orava_optimal_k(n_samples: f64) -> f64 {
 /// The optimal `k` is computed using Orava's formula.
 ///
 /// Returns `None` when `samples` is empty.
-pub fn knn_pdf<X, S>(x: X, samples: &S) -> Result<f64, DensityError>
+pub fn knn_pdf<X, S>(x: X, samples: &S, bandwidth: Option<f64>) -> Result<f64, DensityError>
 where
     S: AsRef<[X]> + Container<Elem = X>,
     X: AsRef<[f64]> + Container + PartialEq,
@@ -46,13 +26,19 @@ where
         return Err(DensityError::EmptySample);
     }
     let n_samples = samples.length() as f64;
-    let k = orava_optimal_k(n_samples);
     let d = x.length();
-    let mut tree = KdTree::with_capacity(d, 1 + n_samples.log2() as usize);
+    let mut tree = KdTree::with_capacity(d, 2usize.pow(n_samples.log2() as u32));
     for (position, sample) in samples.as_ref().iter().enumerate() {
         tree.add(sample.clone(), position)?;
     }
-    let neighbors = tree.nearest(x.as_ref(), k as usize, &squared_euclidean)?;
+    let (neighbors, k) = if let Some(bandwidth) = bandwidth {
+        let neighbors = tree.within(x.as_ref(), bandwidth, &squared_euclidean)?;
+        let k = neighbors.len() as f64;
+        (neighbors, k)
+    } else {
+        let k = orava_optimal_k(n_samples);
+        (tree.nearest(x.as_ref(), k as usize, &squared_euclidean)?, k)
+    };
     if neighbors.is_empty() {
         Err(DensityError::EmptyNeighborhood)
     } else {
@@ -65,8 +51,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::distribution::{Continuous, Normal};
-    use nalgebra::Vector1;
+    use crate::distribution::Normal;
+    use nalgebra::Vector2;
     use rand::distributions::Distribution;
 
     #[test]
@@ -74,12 +60,14 @@ mod tests {
         let law = Normal::new(0., 1.).unwrap();
         let mut rng = rand::thread_rng();
         let samples = (0..10000)
-            .map(|_| Vector1::new(law.sample(&mut rng)))
+            .map(|_| Vector2::new(law.sample(&mut rng), law.sample(&mut rng)))
             .collect::<Vec<_>>();
-        let x = 0.0;
-        let knn_density = knn_pdf(Vector1::new(x), &samples);
+        let x = Vector2::new(0.0, 0.0);
+        let knn_density_with_bandwidth = knn_pdf(x, &samples, Some(0.75));
+        let knn_density = knn_pdf(x, &samples, None);
         println!("Knn: {:?}", knn_density);
-        println!("Pdf: {:?}", law.pdf(x));
+        println!("Knn with bandwidth: {:?}", knn_density_with_bandwidth);
+        // println!("Pdf: {:?}", law.pdf(x));
         assert!(knn_density.is_ok());
     }
 
@@ -87,7 +75,7 @@ mod tests {
     fn test_knn_pdf_empty_samples() {
         let samples: Vec<[f64; 1]> = vec![];
         let x = 3.0;
-        let result = knn_pdf([x], &samples);
+        let result = knn_pdf([x], &samples, None);
         assert!(matches!(result, Err(DensityError::EmptySample)));
     }
 }
